Added interfaces for Service layer
